pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rql"
version = "0.1.0"
description = "RQL (Retrieval Query Language) - A language for AI-powered data retrieval and processing"
authors = [
    {name = "RQL Contributors"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer[all]>=0.12.0",
    "lark>=1.1.8",
    "pydantic>=2.7.4",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "pygments>=2.17.0",
    "toml>=0.10.2",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-httpx>=0.26.0",
    "pytest-snapshot>=0.9.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[project.scripts]
rql = "rql.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["rql"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
